refactor: Create types.ts
