--- app/_types/file.tsx
@@ -1,5 +1,0 @@
-export type File = {
-  key: string,
-  data: number[],
-  contentType: string,
-}
