+++ src/app/mypage/projects/_components/ProjectCard.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { Project } from "@/app/_types/project";
+import { Project } from "@/types";
 import {
   Card,
   CardContent,
