+++ src/app/projects/(search)/_components/ProjectSearchList.tsx
@@ -1,4 +1,4 @@
-import { Project } from '@/src/app/_types/project'
+import { Project } from '@/types'
 import { useRouter } from 'next/navigation'
 import {
   List,
