+++ src/app/projects/(search)/action.ts
@@ -1,7 +1,7 @@
 'use server'
 
-import { Project } from '@/src/app/_types/project'
-import { prisma } from '@/src/prisma'
+import { Project } from '@/types'
+import { prisma } from '@/prisma'
 
 export const action = async (prevState: any, formData: FormData) => {
   const query = formData.get('query')?.toString() ?? ''
