+++ src/app/projects/[projectId]/loader.ts
@@ -1,4 +1,4 @@
-import { prisma } from '@/src/prisma'
+import { prisma } from '@/prisma'
 import { auth } from "../../../auth"
 
 export const getRole = async (projectId: string) => {
