+++ src/app/projects/new/action.ts
@@ -8,7 +8,7 @@ import { redirect } from 'next/navigation'
 import { openai } from '@ai-sdk/openai'
 import { PutObjectCommand, S3Client } from '@aws-sdk/client-s3'
 import { fromCognitoIdentityPool } from '@aws-sdk/credential-providers'
-import { prisma } from '@/src/prisma'
+import { prisma } from '@/prisma'
 import { generateFilesObjectAgent, streamSummaryTextAgent } from './_agent/agents'
 import { documentsLoader, imagesLoader } from './_agent/loader'
 import { auth } from '../../../auth'
